requirements.txt
discord.py==2.3.2
google-generativeai==0.8.3
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
requests==2.31.0